move configs
