+++ python/ray/serve/_private/client.py
@@ -16,7 +16,8 @@ from ray.serve._private.common import (
     DeploymentStatusInfo,
     MultiplexedReplicaInfo,
 )
-from ray.serve.config import DeploymentConfig, HTTPOptions, ReplicaConfig
+from ray.serve._private.config import DeploymentConfig, ReplicaConfig
+from ray.serve.config import HTTPOptions
 from ray.serve._private.constants import (
     CLIENT_POLLING_INTERVAL_S,
     CLIENT_CHECK_CREATION_POLLING_INTERVAL_S,
