+++ python/ray/serve/_private/common.py
@@ -5,7 +5,7 @@ from typing import Any, List, Dict, Optional, NamedTuple
 
 import ray
 from ray.actor import ActorHandle
-from ray.serve.config import DeploymentConfig, ReplicaConfig
+from ray.serve._private.config import DeploymentConfig, ReplicaConfig
 from ray.serve.generated.serve_pb2 import (
     DeploymentInfo as DeploymentInfoProto,
     DeploymentStatusInfo as DeploymentStatusInfoProto,
