+++ python/ray/serve/_private/deployment_function_node.py
@@ -4,7 +4,7 @@ from ray.dag.dag_node import DAGNode
 from ray.dag.format_utils import get_dag_node_str
 
 from ray.serve.deployment import Deployment, schema_to_deployment
-from ray.serve.config import DeploymentConfig, ReplicaConfig
+from ray.serve._private.config import DeploymentConfig, ReplicaConfig
 from ray.serve.handle import DeploymentHandle, RayServeHandle
 from ray.serve.schema import DeploymentSchema
 from ray.serve._private.constants import RAY_SERVE_ENABLE_NEW_HANDLE_API
