+++ python/ray/serve/_private/version.py
@@ -4,9 +4,9 @@ import json
 from typing import Any, Optional, Dict, List
 from zlib import crc32
 
+from ray.serve._private.config import DeploymentConfig
 from ray.serve._private.utils import get_random_letters, DeploymentOptionUpdateType
 from ray.serve.generated.serve_pb2 import DeploymentVersion as DeploymentVersionProto
-from ray.serve.config import DeploymentConfig
 
 import logging
