+++ python/ray/serve/tests/test_deployment_state.py
@@ -8,14 +8,13 @@ import pytest
 import ray
 from ray.serve._private.common import (
     DeploymentID,
-    DeploymentConfig,
     DeploymentInfo,
     DeploymentStatus,
-    _ReplicaConfig,
     ReplicaTag,
     ReplicaName,
     ReplicaState,
 )
+from ray.serve._private.config import DeploymentConfig, ReplicaConfig
 from ray.serve._private.deployment_scheduler import (
     ReplicaSchedulingRequest,
 )
@@ -308,7 +307,7 @@ def deployment_info(
         deployment_config=DeploymentConfig(
             num_replicas=num_replicas, user_config=user_config, **config_opts
         ),
-        replica_config=_ReplicaConfig.create(lambda x: x),
+        replica_config=ReplicaConfig.create(lambda x: x),
         deployer_job_id="",
         is_driver_deployment=is_driver_deployment,
     )
